+++ Sources/Danger/Danger.swift
@@ -50,7 +50,7 @@ final class DangerRunner {
         }
         do {
             let decoder = JSONDecoder()
-            decoder.dateDecodingStrategy = .formatted(DateFormatter.defaultDateFormatter)
+            decoder.dateDecodingStrategy = .custom(DateFormatter.dateFormatterHandler)
             logger.debug("Decoding the DSL into Swift types")
             dsl = try decoder.decode(DSL.self, from: dslJSONContents).danger
         } catch {
@@ -62,7 +62,7 @@ final class DangerRunner {
         outputPath = outputJSONPath
         dumpResults()
     }
-    
+
     private func dumpResults() {
         logger.debug("Sending results back to Danger")
         do {