+++ Sources/Danger/Extensions/DateFormatterExtensions.swift
@@ -6,6 +6,26 @@ extension DateFormatter {
         let dateFormatter = OptionalFractionalSecondsDateFormatter()
         return dateFormatter
     }
+
+    public static var onlyDateDateFormatter: DateFormatter {
+        let dateFormatter = OptionalFractionalSecondsDateFormatter.onlyDate
+        return dateFormatter
+    }
+
+    /// Handles multiple date format inside models.
+    public static func dateFormatterHandler(_ decoder: Decoder) throws -> Date {
+        let dateString = try decoder.singleValueContainer().decode(String.self)
+        if let date = defaultDateFormatter.date(from: dateString) {
+            return date
+        } else if let date = onlyDateDateFormatter.date(from: dateString) {
+            return date
+        } else {
+            let path = decoder.codingPath.map { $0.stringValue }.joined(separator: ".")
+            throw OptionalFractionalSecondsDateFormatter
+                .DateError
+                .invalidFormat(path: path, dateString: dateString)
+        }
+    }
 }
 
 private final class OptionalFractionalSecondsDateFormatter: DateFormatter {
@@ -13,6 +33,13 @@ private final class OptionalFractionalSecondsDateFormatter: DateFormatter {
         let formatter = DateFormatter()
         setUpFormatter(formatter)
         formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSSSSZZZZ"
+        return formatter
+    }()
+
+    static let onlyDate: DateFormatter = {
+        let formatter = DateFormatter()
+        setUpFormatter(formatter)
+        formatter.dateFormat = "yyyy-MM-dd"
         return formatter
     }()
 
@@ -42,5 +69,20 @@ private final class OptionalFractionalSecondsDateFormatter: DateFormatter {
             return result
         }
         return OptionalFractionalSecondsDateFormatter.withMilliseconds.date(from: string)
+    }
+}
+
+extension OptionalFractionalSecondsDateFormatter {
+    enum DateError: LocalizedError {
+        case invalidFormat(path: String, dateString: String)
+
+        // MARK: LocalizedError
+
+        var errorDescription: String? {
+            switch self {
+            case let .invalidFormat(path, dateString):
+                return "Format Invalid with path \"\(path)\", date string: \"\(dateString)\""
+            }
+        }
     }
 }