+++ Tests/DangerDependenciesResolverTests/PackageDataProviderTests.swift
@@ -230,12 +230,12 @@ final class MockedExecutor: ShellExecuting {
     var receivedCommand: String!
     var result = ""
 
-    func execute(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile: String?) -> String {
+    func execute(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile _: String?) -> String {
         receivedCommand = command + " " + arguments.joined(separator: " ")
         return result
     }
 
-    func spawn(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile: String?) throws -> String {
+    func spawn(_ command: String, arguments: [String], environmentVariables _: [String: String], outputFile _: String?) throws -> String {
         receivedCommand = command + " " + arguments.joined(separator: " ")
         return result
     }