+++ Tests/DangerTests/GitDiffTests.swift
@@ -143,6 +143,8 @@ final class GitDiffTests: XCTestCase {
         ]))
     }
 
+    // swiftlint:disable all
+    // swiftformat:disable all
     var testDiff: String {
         """
         diff --git a/.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata b/.swiftpm/xcode/package.xcworkspace/contents.xcworkspacedata